add new database

+++ src/ConferencesManagement/ConferencesManagement/Common/UserLogin.cs
@@ -8,7 +8,7 @@ namespace ConferencesManagement.Common
     [Serializable]
     public class UserLogin
     {
-        public int IdAccount { get; set; }
+        public long IdAccount { get; set; }
         public string UserName { get; set; }
     }
 }
